requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
urllib3==2.0.7
